+++ build.gradle.kts
@@ -1,9 +1,7 @@
 plugins {
     kotlin("multiplatform") apply false
     id("com.vanniktech.maven.publish") version "0.21.0" apply false
-
 }
-
 
 buildscript {
     repositories {
