+++ sandbox/src/jsMain/kotlin/JsMain.kt
@@ -1,0 +1,23 @@
+import com.example.api.JensPlaceHolderApi
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.launch
+
+
+fun main() {
+
+
+
+
+
+    GlobalScope.launch {
+        println("Launch")
+
+        //val api2: JensPlaceHolderApi = jvmKtorfit.create()
+
+        delay(3000)
+
+    }
+
+
+}
